Separate top level component from the container component.
